Test that 'baseline' pings are sent by Fenix
This is an initial instrumented test for Fenix
that checks if a 'baseline' ping is generated when
going to background. This is a bit tricky, as
it requires to customise the test runner to
spin up an http server to receive the ping submissions
and that can't happen on the main thread.